Add side-nav to manage page
