+++ frontend/src/App.js
@@ -1,7 +1,14 @@
 import React, { useState, useEffect } from "react";
 import { Routes, Route, Outlet, Navigate } from "react-router-dom";
+import ScrollContainer from "react-indiana-drag-scroll";
 import { token_clear, api_validtoken } from "./auth";
-import { SITE_PAGES, SIDENAV_STEPS, SIDENAV_ROUTES } from "./constants/links";
+import {
+  SITE_PAGES,
+  SIDENAV_STEPS,
+  SIDENAV_ROUTES,
+  MANAGE_STEPS,
+  MANAGE_ROUTES,
+} from "./constants/links";
 import PageLayout from "./components/PageLayout";
 import LoginPage from "./pages/LoginPage";
 import ProfilePage from "./pages/ProfilePage";
@@ -9,6 +16,9 @@ import WishGrantingPage from "./pages/WishGrantingPage";
 import Custom404Page from "./pages/Custom404Page";
 import ManagePage from "./pages/ManagePage";
 import WishStep from "./components/WishStep";
+
+import UserList from "./components/UserList";
+import AssignBtn from "./components/AssignBtn";
 
 function ProtectedRoute({ isAuth, setIsAuth }) {
   const [hasFired, setHasFired] = useState(false);
@@ -33,6 +43,135 @@ function SignoutHelper({ setIsAuth }) {
 
   return <Navigate to={SITE_PAGES.LOGIN} />;
 }
+
+function ButtonContainer({ btnLabels }) {
+  return (
+    <ScrollContainer className="assign_btn_container" vertical={false}>
+      {btnLabels.map((label) => (
+        <AssignBtn label={label} key={Math.random()} />
+      ))}
+    </ScrollContainer>
+  );
+}
+
+// NOTE: This is just a temporary implementation for the MVP
+function getUserData() {
+  return [
+    {
+      Name: "Admin 1 (Carly)",
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      Admin: true,
+      // Completed: 3,
+      // Start: "May 2017",
+    },
+    {
+      Name: "Admin 2 (Gibson)",
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      Admin: true,
+      // Roles: [<ButtonContainer btnLabels={["Airport Greeter", "Wish Granter"]} />],
+      // Completed: 1,
+      // Start: "Feb 2019",
+    },
+    {
+      Name: "Admin 3 (Freddie)",
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      Admin: true,
+      // Roles: [<ButtonContainer btnLabels={["Wish Granter", "Meeter"]} />],
+      // Completed: 2,
+      // Start: "June 2016",
+    },
+    {
+      Name: "Admin 4 (Bob)",
+      Admin: true,
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      // Roles: [<ButtonContainer btnLabels={["Wish Granter"]} />],
+      // Completed: 2,
+      // Start: "May 2020",
+    },
+    {
+      Name: "Volunteer 1 (Bob)",
+      Admin: false,
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      // Roles: [<ButtonContainer btnLabels={["Wish Granter", "Airport Greeter"]} />],
+      // Completed: 2,
+      // Start: "June 2020",
+    },
+    {
+      Name: "Volunteer 2 (Rob)",
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      Admin: false,
+      // Roles: [<ButtonContainer btnLabels={["Wish Granter", "Airport Greeter"]} />],
+      // Completed: 2,
+      // Start: "June 2020",
+    },
+    {
+      Name: "Volunteer 3 (Freddie)",
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      Admin: false,
+      // Roles: [<ButtonContainer btnLabels={["Wish Granter", "Airport Greeter"]} />],
+      // Completed: 2,
+      // Start: "June 2020",
+    },
+    {
+      Name: "Volunteer 4 (Rib)",
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      Admin: false,
+      // Roles: [<ButtonContainer btnLabels={["Wish Granter", "Airport Greeter"]} />],
+      // Completed: 2,
+      // Start: "June 2020",
+    },
+    {
+      Name: "Volunteer 5 (Pete)",
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      Admin: false,
+      // Roles: [<ButtonContainer btnLabels={["Wish Granter", "Airport Greeter"]} />],
+      // Completed: 2,
+      // Start: "June 2020",
+    },
+    {
+      Name: "Volunteer 6 (Alice)",
+      Roles: [<ButtonContainer btnLabels={["Allow Access"]} />],
+      Admin: false,
+      // Roles: [<ButtonContainer btnLabels={["Wish Granter", "Airport Greeter"]} />],
+      // Completed: 2,
+      // Start: "June 2020",
+    },
+  ];
+}
+
+const userData = getUserData();
+
+const headers = [
+  {
+    Header: "Name",
+    accessor: "Name",
+  },
+  // Replace the following three rows with the commented out rows for the full table
+  {
+    Header: "",
+    accessor: "Roles",
+  },
+  {
+    Header: "",
+    accessor: "empty",
+  },
+  {
+    Header: "",
+    accessor: "empty1",
+  },
+  // {
+  //   Header: "Roles",
+  //   accessor: "Roles",
+  // },
+  // {
+  //   Header: "Assignments Completed",
+  //   accessor: "Completed",
+  // },
+  // {
+  //   Header: "Volunteer Since",
+  //   accessor: "Start",
+  // },
+];
 
 function App() {
   const [isAuth, setIsAuth] = useState(false);
@@ -62,7 +201,21 @@ function App() {
               <ManagePage />
             </PageLayout>
           }
-        />
+        >
+          <Route
+            exact
+            path={SITE_PAGES.MANAGE}
+            element={<Navigate to={`${SITE_PAGES.MANAGE}/${MANAGE_ROUTES[0]}`} />}
+          />
+          {MANAGE_STEPS.map((name, ind) => (
+            <Route
+              exact
+              key={name}
+              path={MANAGE_ROUTES[ind]}
+              element={<UserList tableHeaders={headers} userData={userData} />}
+            />
+          ))}
+        </Route>
         {/* Redirect to Manage Page, only when authenticated */}
         <Route exact path="/" element={<Navigate to={SITE_PAGES.MANAGE} />} />
         {/* Wish Granting Page */}
